requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-cohere
langchain-google-genai
chromadb
python-dotenv
pydantic
langdetect
requests
pypdf
